backend/requirements.txt
django==1.11
gunicorn==19.7
django_extensions==1.6.7
django-localflavor==1.4
djangorestframework==3.6
pillow==2.7.0
six==1.9.0

ipdb
werkzeug==0.10.1
